test: spec for text content decoder and basic auth added
